Create a column definition for use with the create column method for SharePoint lists

create_column_definition() builds a named list with the properties of the columnDefinition resource type.

Usage

create_column_definition(
  name,
  ...,
  .col_type = "text",
  enforce_unique = NULL,
  hidden = FALSE,
  deletable = NULL,
  indexed = NULL,
  sealed = NULL,
  propagate_changes = NULL,
  read_only = NULL,
  required = NULL,
  validation = NULL,
  default = get_column_default(),
  description = NULL,
  displayname = NULL,
  id = NULL
)

create_text_column(
  name,
  ...,
  multiple_lines = NULL,
  append_changes = NULL,
  lines = NULL,
  max_length = NULL,
  text_type = c("plain", "richText")
)

create_choice_column(
  name,
  choices,
  ...,
  allow_text = TRUE,
  display_as = c("checkBoxes", "dropDownMenu", "radioButtons"),
  allow_na = TRUE,
  na_replacement = "NA",
  split = NULL
)

create_number_column(
  name,
  ...,
  decimals = "automatic",
  display_as = NULL,
  max = NULL,
  min = NULL
)

create_datetime_column(
  name,
  ...,
  display_as = c("default", "friendly", "standard"),
  format = c("dateOnly", "dateTime")
)

create_boolean_column(name, ...)

create_currency_column(name, ..., locale = "en-us")

create_calculated_column(
  name,
  ...,
  formula,
  format = c("dateOnly", "dateTime"),
  output_type = c("boolean", "currency", "dateTime", "number", "text")
)

create_lookup_column(
  name,
  lookup_list_column,
  ...,
  lookup_list_id = NULL,
  lookup_list = NULL,
  allow_multiple = NULL,
  allow_unlimited_length = NULL,
  primary_lookup_column_id = NULL
)

create_person_column(
  name,
  ...,
  allow_multiple = NULL,
  display_as = NULL,
  from_type = "peopleOnly"
)

create_group_column(
  name,
  ...,
  allow_multiple = NULL,
  display_as = NULL,
  from_type = "peopleAndGroups"
)

create_hyperlink_column(name, ..., is_picture = FALSE)

create_picture_column(name, ..., is_picture = TRUE)

create_thumbnail_column(name, ...)

create_geolocation_column(name, ...)

Arguments

format

"dateOnly" or "dateTime". Required if output_type is "dateTime" otherwise ignored.

formula

Required string with formula for calculated column definition.

output_type

Value type returned by calculated formula.

lookup_list_column

Name of lookup column in the lookup list to use.

lookup_list_id, lookup_list

Lookup list ID string or "ms_list" class object with id value in list properties.

allow_multiple

If TRUE, allow lookup column to return multiple values.

allow_unlimited_length

If TRUE, allow lookup column to return any length value.

primary_lookup_column_id

If column definnition is for a secondary column, the primary lookup column ID must be supplied.

from_type

What type of resources to choose from. Defaults to "peopleOnly" for create_person_column() or "peopleAndGroups" for create_group_column()

is_picture

Logical indicator for display of hyperlink value as link (FALSE, default for create_hyperlink_column()) or image (TRUE, default for create_picture_column()).

Details

https://learn.microsoft.com/en-us/graph/api/resources/columndefinition?view=graph-rest-1.0

Display as options

Display as options vary by columnDefinition type:

Options for personOrGroupColumn: https://learn.microsoft.com/en-us/graph/api/resources/personorgroupcolumn?view=graph-rest-1.0#displayas-options