Read a SharePoint item based on a file URL or file name and site details

read_sharepoint() is designed to download a SharePoint item to a temporary folder and read the file based on the file extension. The class of the returned object depends on the file extension of the file parameter.

If the file is a csv, csv2, or tsv file or a Microsoft Excel file (xlsx or xls), readr::read_delim() or readxl::read_excel() is used to return a data frame.

If the file is a rds file, readr::read_rds() is used to return the saved object.

If the file is a pptx or docx file, officer::read_docx() or officer::read_pptx() are used to read the file into a rdocx or rpptx object.

If the file has a "gpkg", "geojson", "kml", "gdb", or "shp" file extension, sf::read_sf() is used to read the file into a sf data frame.

If the file has none of these file extensions, an attempt is made to read the file with readr::read_lines().

Usage

read_sharepoint(
  file,
  ...,
  new_path = tempdir(),
  drive_name = NULL,
  drive_id = NULL,
  drive = NULL,
  site_url = NULL,
  site_name = NULL,
  site_id = NULL,
  site = NULL,
  dest = NULL,
  overwrite = TRUE,
  recursive = FALSE,
  parallel = FALSE
)

Arguments

file

Required. A SharePoint shared file URL, document URL, or, if item_id is supplied, a file name to use in combination with new_path to set dest with location and filename for downloaded item.

...

Additional parameters passed to one of the functions identified in the description.

new_path

Path to directory for downloaded item. Optional if dest is supplied.

drive_name, drive_id

SharePoint drive name or ID.

drive

A ms_drive object. If drive is supplied, drive_name, site_url, and any additional parameters passed to ... are ignored.

site_url

A SharePoint site URL in the format "https://[tenant name].sharepoint.com/sites/[site name]". Any SharePoint item or document URL will be parsed and a site URL built using the tenant and site name if found.

site_name, site_id

Site name or ID of the SharePoint site as an alternative to the SharePoint site URL. Only one of the three arguments should be supplied.

site

A ms_site object. If site is supplied, site_url is ignored.

dest, overwrite, recursive, parallel

Parameters passed to download method for ms_drive_item object.